refactor: centralize applyCors utility
