+++ shared/utils/applyCors.ts
