+++ smoothr/pages/api/create-checkout-session.ts
@@ -2,7 +2,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import Stripe from 'stripe';
 import { createServerSupabaseClient } from 'shared/supabase/serverClient';
 import { getStoreIntegration } from 'shared/checkout/getStoreIntegration';
-import { applyCors } from '../../utils/cors';
+import { applyCors } from '../../../shared/utils/applyCors';
 
 const debug = process.env.SMOOTHR_DEBUG === 'true';
 const log = (...args: any[]) => debug && console.log('[create-checkout]', ...args);
