+++ smoothr/pages/api/get-payment-key.js
@@ -1,6 +1,6 @@
 // pages/api/get-payment-key.js
 import { createClient } from '@supabase/supabase-js';
-import { applyCors } from '../../utils/cors';
+import { applyCors } from '../../../shared/utils/applyCors';
 
 export default async function handler(req, res) {
   const origin = process.env.CORS_ORIGIN || '*';
